Added primitive combat
Combat added. Health declines successfully for both parties.

A primitive death system works for the character.

Having trouble getting enemy death to work.
When the enemy health crosses 0, I want to add text to the next prompt
or add a special prompt line?
I also want to remove the instance of that monster, having trouble with
that as well. Possibly the way I generate the 'special' dictionary makes
that difficult? Would it be easier to remove the enemy by having the
special generator only run once, then on death removing the reference of
the monster in special?